FlaskAPI/requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.1.2
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.23.3
pandas==1.5.0
plotly==5.8.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.1.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
